Trine (35700), and update test to disable Plat check
It seems that each game I decide to look at will have another
problematic behaviour! Awesomely this game includes a Platinum-like
trophy on Steam, but it also has unmappable achievements and trophies!
As well as having requirements that differ (depending on release order,
they were changed to be less frustrating) and yet they added some harder
ones on Steam too. Assuming I have to revisit this in the future with
people complaining the ‘All Boxes And No Play…’ (500 vs 300) and
‘Whoops!’ (3 vs 5) have differences enough that it is unfair to map.

We also bump into Trine:EE being an update on Steam, but a different
title on PSN (Trine on PS3, Trine: EE on PS4) so I have to look at that
next!