Grand Theft Auto V (NPWR06221_00→271590)
This is the PS4 edition, note the extra trophy for First Person mode.
Because of the extra trophy, this cannot cleanly map with a duplicate
flag, so we relist the mappings including the extra achievement.
Need to look into DLC behaviour a bit more and then decide the best way
to store mapping