Update README to cover zero-indexed mappings
Probably failed to describe this well in the README, but some titles on
PSN do not have platinums. Which would be fine in an ideal world, but
some games like to add this meta‐completion trophy as a Steam
achievement, so we need to cover that mapping. It means that some
contain "0" due to mapping the Platinum, and the rest because they do
not have one at all. Wish there were a better way to do that (ie, would
like to expose if a title has Platinum)