Publishing my TODO thoughts as an RFC
We need to describe and delimit dlc/main trophy brackets, allow for gaps
or voids with unmapped trophies, have some support for fallback so we
can get away with *not* being explicit, have some flexibility for ranges
and mapping when we need it. Still need to show a use case for the
paired achievement mapping.

Length and number order probably need more fleshing out, so we should
make a simple tool to enumerate the mapping from a brief description.