Call of Duty®: Black Ops needs some adjustment
OK, this is messy, and I don’t really know a good solution to it.
Basically this game is torn in two, with single‐player and multi‐player
being different appids. Which would be complex enough, without Aspyr
being forced to make *new* appids for their OSX ports. Oh, and their
MP app has one fewer achievement. For no good damn reason.

So 4 appids mapped to 1 PSN game, but some achievements have to be
triggered on the specific appid that includes them. So yeah. This might
need adjustments to my new design already. Glad I did not push that yet…