Make mapped trophy numbers implicit
Started working on the new mapping format, this change flattens
achievement lists so we no longer have to worry about explicit
numbering. This is meant to ease updates, and also make some additional
‘compression’ tweaks for the storage format give us easy wins.

This does however make platinums rather ambiguous, as now we just set ""
if it has one - I would rather prefer us to say, this is a platinum
somehow. WIP.