Update BioShock 2 to include DLC
This exposes more functionality that I wanted to add for the schema
update, offset can now use a local variable (group) for group‒local
counter increments. This is reset to 0 upon entering a new group and
increments one per loop, so in this case group+200 would be 200, 201,
202 and so on.
Another thing to mention, some DLC have appids on Steam, and some do
not. With how Steam appears to have changed over the years, we are going
to track appids probably when we do not need to, as there are some cases
where it _is_ needed. Yay.