Add additional appid for Batman: Arkham Asylum
This was bothering and boggling me. Steam supports migrating games onto
new appids, by giving the user access to the new appid alongside the
previous, and then… well, that is about it.
Batman: AA for the re‐released GOTY Edition had support for Steam
Achievements added. So achievements need to be triggered on 35140.
However we do not yet have a way of marking that in our dataset.