Update Batman: Arkham City™ to include DLC
Gosh, this mapping is a mess, huh? Main game has a bifurcated range of
NEW_ACHIEVEMENT_1_* and NEW_ACHIEVEMENT_2_* achievements, some DLC is
integrated on Steam whereas others are not. GOTY Edition merges all of
the DLC into the main package (so one will have to be aware that a query
for ownership may fail for those DLC owners?). And there is no
discernable mark between DLC in achievement names, as it seems
internally the game uses 0-32 numbering for each range.

This demonstrates the overlapping feature of groups, as the first two
groups are part of the main game (they have no name set, so it fallbacks
onto ‘base’, merging the two entries). Which could also have been used
for the Harley PDLC group as it contains two possible variable expansion
mappings. But in some cases, it is easier to just provide a list!

This does unmap some known missing achievements on Steam, and I would
still like to support this somehow. Maybe reintroducing explicit mapping
as a following group to override any variable mappings? i.e.

	{
		"map":"NEW_ACHIEVEMENT_2_%d",
		"offset":"group",
		"length":18
	},
	{
		"map":{
			"15":"-1",
			"19":"-1",
			"21":"-1"
		}
	}