Update Call of Duty®: Black Ops to include DLC
This entry was a mess. It is now tidier for the base game. And a bigger
mess for the remaining DLC.
At least bundling appids together for the Win+OSX DLC makes it more
simple for the dataset.

Here we show off attaching arbitary data onto the middle of the trophy
listing for the 3 multiplayer trophies. Platinum, group of single player
trophies, group of mp, a few more sp, then the DLC groups. Not setting
the "name" in this group of 3 MP trophies makes it work, otherwise they
would not apply to the base game. Maybe? This is complicated.