Update Call of Duty®: Black Ops II to include DLC
Now this is an interesting case I have not come across yet, various DLC
achievements are per of a sub appid for Zombies: 212910. Our current
system only infers dlc ownership with the extra appid metadata, not the
parentage of the achievements. So it is likely in a future update we
need to adjust this further, after actual interaction with the API

Future John: these DLCs need something like "parent":212910, to suggest
on what appid to trigger the achievements. Or instead, we rename groups
appid to "ownership" or something.